requirements.txt
beautifulsoup4==4.11.1
google_api_python_client==2.72.0
joblib==1.2.0
moviepy==1.0.3
nltk==3.8.1
numpy==1.24.1
pandas==1.5.2
pydub==0.25.1
PyQt5==5.15.7
python-dotenv==0.21.0
pytube==12.1.0
requests==2.28.1
scikit_learn==1.2.0
SpeechRecognition==3.8.1
PyQtChart==5.15.6
selenium==4.7.2
webdriver-manager==3.8.5